fix: address review feedback — constants.py, types.py, click CLI
Splits scripts/dispatch.py into a scripts/dispatch/ package so the module
layout follows the house rules the flat file was breaking:

- constants.py — the 11 Final constants, now shared and public-named
  (python.md, matching installer/constants.py and metrics/constants.py)
- types.py — DispatchError, the four frozen dataclasses, CommandRunner
- plan.py / windows.py — parsing at the boundary, derivation + execution
- cli.py — click command, subprocess runner, resolvers

A package rather than bare scripts/constants.py + scripts/types.py: that
directory is staged flat into ~/.claude/at/scripts/ alongside other packages'
scripts, and a top-level types.py would shadow the stdlib `types` module once
its directory lands on sys.path[0].

argparse -> click, the library installer/at.py already standardises on. The
command keeps a plain run_cli() seam so the existing tests drive it directly;
two CliRunner tests cover the decorator wiring itself.

Invocation moves to `PYTHONPATH=~/.claude/at/scripts uv run --no-project
--with click python -m dispatch`, following the same uv-run-with-dep pattern
make-pr already uses for validate_return.py. Verified end-to-end against a
staged copy of the package: worktrees, windows, and launch all correct.
