+++ scripts/dispatch/__main__.py
@@ -1,0 +1,8 @@
+"""Entry point for `python -m dispatch`."""
+
+from __future__ import annotations
+
+from .cli import cli
+
+if __name__ == "__main__":
+    cli()
